+++ skillfarm/models/general.py
@@ -1,12 +1,22 @@
 """Model for Permission."""
 
+# Standard Library
+from dataclasses import dataclass
+from typing import Any, NamedTuple
+
 # Django
 from django.db import models
 
-# AA Skillfarm
-from skillfarm.hooks import get_extension_logger
+# Alliance Auth
+from allianceauth.services.hooks import get_extension_logger
 
-logger = get_extension_logger(__name__)
+# Alliance Auth (External Libs)
+from app_utils.logging import LoggerAddTag
+
+# AA Skillfarm
+from skillfarm import __title__
+
+logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 class General(models.Model):
@@ -20,3 +30,26 @@ class General(models.Model):
             ("corp_access", "Has access to all characters in the corporation."),
             ("admin_access", "Has access to all characters."),
         )
+
+
+class UpdateSectionResult(NamedTuple):
+    """A result of an attempted section update."""
+
+    is_changed: bool | None
+    is_updated: bool
+    data: Any = None
+
+
+@dataclass(frozen=True)
+class _NeedsUpdate:
+    """An Object to track if an update is needed."""
+
+    section_map: dict[str, bool]
+
+    def __bool__(self) -> bool:
+        """Check if any section needs an update."""
+        return any(self.section_map.values())
+
+    def for_section(self, section: str) -> bool:
+        """Check if an update is needed for a specific section."""
+        return self.section_map.get(section, False)