+++ skillfarm/templatetags/skillfarm.py
@@ -7,12 +7,17 @@ from django.template.defaulttags import register
 from django.templatetags.static import static
 from django.utils.safestring import mark_safe
 
+# Alliance Auth
+from allianceauth.services.hooks import get_extension_logger
+
+# Alliance Auth (External Libs)
+from app_utils.logging import LoggerAddTag
+
 # AA Skillfarm
 from skillfarm import __title__, __version__
 from skillfarm.helpers.static_files import calculate_integrity_hash
-from skillfarm.hooks import get_extension_logger
 
-logger = get_extension_logger(__name__)
+logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 @register.simple_tag