+++ skillfarm/urls.py
@@ -12,6 +12,7 @@ app_name: str = "skillfarm"
 urlpatterns = [
     # -- Views
     path("", views.index, name="index"),
+    path("admin/", views.admin, name="admin"),
     path(
         "<int:character_id>/view/skillfarm/",
         views.skillfarm,